requirements.txt
streamlit>=1.28.0
requests>=2.31.0

# 后端依赖（管理后台需要）
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv==1.0.0

# 文档处理
python-docx>=0.8.11
Pillow>=9.0.0
lxml>=4.9.0
